+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Minor: Added `ban` method to `ChatClient` API (#186).
+
 ## v4.2.0
 
 - Bugfix: Fixed typescript compiling error (#154)